Move all data to one file, since these will come from backend later on
