+++ frontend/src/allData.js
@@ -1,0 +1,121 @@
+export const workhistory = [
+  {
+    position: 'Software engineer',
+    employer: 'Vincit Oyj',
+    duration: '06/2018 – present',
+  },
+  {
+    position: 'Software summer trainee',
+    employer: 'Nordic Semiconductor ASA',
+    duration: '05/2017 – 08/2018',
+  },
+  {
+    position: 'Goalkeeper coach',
+    employer: 'Tervarit juniorit ry / OLS',
+    duration: '11/2015 – 06/2017'
+  },
+  {
+    position: 'Sales negotiatior',
+    employer: 'Thinko Positivo Oy',
+    duration: '11/2015 – 12/2016',
+  },
+  {
+    position: 'Sales negotiatior',
+    employer: 'ProCenter Yhteyskeskus Oy',
+    duration: '05/2015 – 08/2015',
+  },
+  {
+    position: 'Second Lieutenant',
+    employer: 'Finnish Defence Forces',
+    duration: '09/2014 – 12/2014',
+  },
+  {
+    position: 'Temporary manpower',
+    employer: 'Nordic Henkilöstöpalvelut',
+    duration: '12/2012 – 08/2013',
+  },
+];
+
+export const languages = [
+  {
+    language: 'Finnish',
+    level: 'Native'
+  },
+  {
+    language: 'English',
+    level: 'Excellent'
+  },
+  {
+    language: 'Swedish',
+    level: 'Good'
+  },
+  {
+    language: 'German',
+    level: 'Good'
+  },
+  {
+    language: 'Latin',
+    level: 'Basics'
+  },
+]
+
+export const education = [
+  {
+    name: 'Master of Science in Wireless Communication',
+    place: 'University of Oulu',
+    duration: '10/2017 – 12/2019',
+  },
+  {
+    name: 'Bachelor of Science in Electrical engineering',
+    place: 'University of Oulu',
+    duration: '08/2013 – 10/2017',
+  },
+  {
+    name: 'Military service',
+    place: 'Finnish defence forces',
+    duration: '2014',
+  },
+  {
+    name: 'Baccalaureate',
+    place: 'Kaleva high school, Tampere',
+    duration: '08/2009 – 05/2013',
+  }
+];
+
+export const positions = [
+  {
+    position: '1st work safety representative',
+    duration: '2019 – 2020',
+    organisation: 'Vincit Oyj',
+  },
+  {
+    position: 'Member of the student union council',
+    duration: '2018 – 2020',
+    organisation: 'University of Oulu',
+  },
+  {
+    position: 'Member of the committee of education',
+    duration: '2018 – 2020',
+    organisation: 'Faculty of information technology and electrical engineering of University of Oulu',
+  },
+  {
+    position: 'Teekkarivastaava',
+    duration: '2017',
+    organisation: 'Sähköinsinöörikilta in university of Oulu',
+  },
+  {
+    position: 'Kiltamestari',
+    duration: '2015',
+    organisation: 'Sähköinsinöörikilta in university of Oulu',
+  },
+  {
+    position: 'Tutor for freshmen',
+    duration: '2015-2016',
+    organisation: 'Sähköinsinöörikilta in university of Oulu',
+  },
+  {
+    position: 'Person in charge of sport activities',
+    duration: '2015',
+    organisation: 'Sähköinsinöörikilta in university of Oulu',
+  },
+]
