--- frontend/src/education.js
@@ -1,22 +1,0 @@
-export const education = [
-  {
-    name: 'Master of Science in Wireless Communication',
-    place: 'University of Oulu',
-    duration: '10/2017 – 12/2019',
-  },
-  {
-    name: 'Bachelor of Science in Electrical engineering',
-    place: 'University of Oulu',
-    duration: '08/2013 – 10/2017',
-  },
-  {
-    name: 'Military service',
-    place: 'Finnish defence forces',
-    duration: '2014',
-  },
-  {
-    name: 'Baccalaureate',
-    place: 'Kaleva high school, Tampere',
-    duration: '08/2009 – 05/2013',
-  }
-];
