--- frontend/src/languages.js
@@ -1,22 +1,0 @@
-export const languages = [
-  {
-    language: 'Finnish',
-    level: 'Native'
-  },
-  {
-    language: 'English',
-    level: 'Excellent'
-  },
-  {
-    language: 'Swedish',
-    level: 'Good'
-  },
-  {
-    language: 'German',
-    level: 'Good'
-  },
-  {
-    language: 'Latin',
-    level: 'Basics'
-  },
-]
