--- frontend/src/positions.js
@@ -1,37 +1,0 @@
-export const positions = [
-  {
-    position: '1st work safety representative',
-    duration: '2019 – 2020',
-    organisation: 'Vincit Oyj',
-  },
-  {
-    position: 'Member of the student union council',
-    duration: '2018 – 2020',
-    organisation: 'University of Oulu',
-  },
-  {
-    position: 'Member of the committee of education',
-    duration: '2018 – 2020',
-    organisation: 'Faculty of information technology and electrical engineering of University of Oulu',
-  },
-  {
-    position: 'Teekkarivastaava',
-    duration: '2017',
-    organisation: 'Sähköinsinöörikilta in university of Oulu',
-  },
-  {
-    position: 'Kiltamestari',
-    duration: '2015',
-    organisation: 'Sähköinsinöörikilta in university of Oulu',
-  },
-  {
-    position: 'Tutor for freshmen',
-    duration: '2015-2016',
-    organisation: 'Sähköinsinöörikilta in university of Oulu',
-  },
-  {
-    position: 'Person in charge of sport activities',
-    duration: '2015',
-    organisation: 'Sähköinsinöörikilta in university of Oulu',
-  },
-]
