+++ frontend/src/screens/CV.tsx
@@ -3,10 +3,7 @@ import logo from '../logo.svg';
 import '../App.css';
 import {Container, Col, Row} from 'react-grid-system';
 
-import {education} from '../education';
-import {positions} from '../positions';
-import {languages} from '../languages';
-import {workhistory} from '../workhistory';
+import {education, positions, languages, workhistory} from '../allData';
 import {DARK_HIGHLIGHT, TEXT_COLOR} from '../assets/constants';
 
 const CV = () => {
@@ -90,6 +87,9 @@ const CV = () => {
         <Row align='start' justify='center'>
           <Col sm={9} md={5} lg={5} className='first' style={styles.colStyle}>
             <h2>My areas of expertise</h2>
+            <p style={{textAlign: 'start'}}>
+              I know stuff<br />And some more
+            </p>
           </Col>
           <Col sm={9} md={5} lg={5} className='second' style={styles.colStyle}>
             <h2>My future plans</h2>
@@ -110,8 +110,5 @@ const styles = {
     backgroundColor: DARK_HIGHLIGHT,
     margin: 15,
     color: TEXT_COLOR,
-  },
-  headerStyle: {
-    textAlign: 'start',
   },
 };
