--- frontend/src/workhistory.js
@@ -1,37 +1,0 @@
-export const workhistory = [
-  {
-    position: 'Software engineer',
-    employer: 'Vincit Oyj',
-    duration: '06/2018 – present',
-  },
-  {
-    position: 'Software summer trainee',
-    employer: 'Nordic Semiconductor ASA',
-    duration: '05/2017 – 08/2018',
-  },
-  {
-    position: 'Goalkeeper coach',
-    employer: 'Tervarit juniorit ry / OLS',
-    duration: '11/2015 – 06/2017'
-  },
-  {
-    position: 'Sales negotiatior',
-    employer: 'Thinko Positivo Oy',
-    duration: '11/2015 – 12/2016',
-  },
-  {
-    position: 'Sales negotiatior',
-    employer: 'ProCenter Yhteyskeskus Oy',
-    duration: '05/2015 – 08/2015',
-  },
-  {
-    position: 'Second Lieutenant',
-    employer: 'Finnish Defence Forces',
-    duration: '09/2014 – 12/2014',
-  },
-  {
-    position: 'Temporary manpower',
-    employer: 'Nordic Henkilöstöpalvelut',
-    duration: '12/2012 – 08/2013',
-  },
-];
